+++ lib/multi_split_view.dart
@@ -1,4 +1,6 @@
 library multi_split_view;
 
 export 'src/multi_split_view.dart';
+export 'src/theme_widget.dart';
+export 'src/theme_data.dart';
 export 'src/controller.dart';